ANDROID: cpufreq: uid_concurrent_policy_time
In order to model the energy used by the cpu,
we need to expose cpu time information to
userspace. Output the time each uid spent running
on group of cpu's with the same policy with 1
to num_related_cpus online and not idle.

The time should also include the time spent by
pids that are no longer alive.

We can use this information to blame uids for
the energy they are using.

For instance, if 2 cores that are a part of the
same policy are online and active for 50ms and
the uid is running on one of the cores, the uid
should be blamed for 1/2 of the energy that is
used everytime one or more cpus in the same policy
is active. Using other proc files, the uid can
also be blamed for the cost of keeping the entire
cpu unsuspended and the cost of the individual
cores it is using.

b/68399339

Test: cat /proc/uid_concurrent_policy_time

Change-Id: I1a5439cdd922ec7edf6fee6a1921fa4561b04e28
Signed-off-by: Marissa Wall <marissaw@google.com>
Signed-off-by: celtare21 <celtare21@gmail.com>